refactor: 패키지 이름 수정, 위치 변경
- common -> global
- move SecurityConfig to /user/auth
