+++ src/main/java/com/pickkasso/pickkasso/global/controller/HomeController.java
@@ -1,4 +1,4 @@
-package com.pickkasso.pickkasso.controller;
+package com.pickkasso.pickkasso.global.controller;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
