+++ src/main/java/com/pickkasso/pickkasso/global/tag/Tag.java
@@ -1,4 +1,4 @@
-package com.pickkasso.pickkasso.common.tag;
+package com.pickkasso.pickkasso.global.tag;
 
 import jakarta.persistence.*;
 import lombok.AccessLevel;
